// src/app/features/models/prices.model.ts
export interface PreciosKeys {
    AltaRFC: 'arfc';
    CFE: 'cfe';
    Cotizadas: 'cot';
    CURP: 'curp';
    Defuncion: 'def';
    VijDerechos: 'der';
    Divorcio: 'div';
    DigiActas: 'dnac';
    EstadoCuentaInfo: 'ecu';
    ConstaInahabilitacion: 'inh';
    ActaMatrimonio: 'mat';
    ActadeNacimiento: 'nac';
}
export enum EstadosKeys {
    BAJACALIFORNIA = 'bcn',
    YUCATAN = 'yuca' ,
    BAJACALIFORNIASUR ='bcs' ,
    VERACRUZ = 'vera',
    COAHUILA = 'coah',
    MICHOACAN = 'mich',
    TLAXCALA = 'tlax',
    DURANGO = 'dura',
    AGUASCALIENTES = 'agua',
    CHIAPAS = 'chia',
    HIDALGO = 'hida',
    PUEBLA = 'pueb',
    QUERETARO = 'quer',
    CHIHUAHA = 'chih',
    OAXACA = 'oaxa',
    SONORA = 'sono',
    SANLUISPOTOSI = 'slp',
    SINALOA = 'sina',
    GUERRERO = 'guer' ,
    ZACATECAS = 'zaca',
    TAMAULIPAS = 'tama',
    MORELOS = 'more',
    TABASCO = 'taba',
    GUANAJUATO = 'guan',
    COLIMA = 'coli',
    JALISCO = 'jali',
    CDMX = 'cdmx',
    NAYARIT = 'naya',
    CAMPECHE = 'camp',
    NUEVOLEON = 'nl',
    MEXICO = 'mex',
    QUINTANAROO = 'qroo',
    EXTRANJERO = 'ext',
}